requirements.txt
torch
torchvision
torchaudio
openai-whisper
tensorboard
